% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/last_2_digits.R
\name{autoplot.last2}
\alias{autoplot.last2}
\title{autoplot last2}
\usage{
\method{autoplot}{last2}(x, ci = 0.95)
}
\arguments{
\item{x}{last2 object}

\item{ci}{quantile of null distribution to plot.}
}
\description{
autoplot function for result of last2_diff_test
}
