% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dup_extract.R
\name{dup_extract}
\alias{dup_extract}
\title{extract duplicate sequences}
\usage{
dup_extract(d, vec, n)
}
\arguments{
\item{d}{result of \code{dup_dist}}

\item{vec}{vector with possible duplicates}

\item{n}{length of sequence sought}
}
\description{
extract duplicate sequences
}
\examples{
#' set.seed(42)
x <- rnorm(10)
x <- c(x, x[1:4] + 2.03)
d <- detectduplicate:::dup_dist(vec = x, n = 3)
detectduplicate:::dup_extract(d, vec = x)
}
