% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dup_find_all.R
\name{dup_find_all}
\alias{dup_find_all}
\title{Find all duplicate sequences}
\usage{
dup_find_all(x, min = 4, tolerance)
}
\arguments{
\item{x}{vector or data.frame with possible duplicates}

\item{min}{length of the shortest sequence of interest
(high risk of false positives if this is short)}

\item{tolerance}{Small positive number to allow for numerical imprecision.
Defaults to \code{sqrt(.Machine$double.eps)}.}
}
\description{
Finds duplicate sequences, identical or with an offset and/or
multiplier, whether forward or reverse.
}
\examples{
data(kp2014)
dup_find_all(
  x = kp2014$`Theridion murarium Aggressiveness...4`,
)

dup_find_all(
  x = kp2014[, c(
    "Theridion murarium Aggressiveness...4",
    "Theridion murarium Aggressiveness...5",
    "Theridion murarium Aggressiveness...6"
  )]
)

}
