% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dup_find_length_n.R
\name{dup_find_length_n}
\alias{dup_find_length_n}
\title{Find and extract duplicate sequences of specified length}
\usage{
dup_find_length_n(vec, n, type, tolerance, reverse)
}
\arguments{
\item{vec}{vector with possible duplicates}

\item{n}{length of sequence sought}

\item{type}{Type of duplicate sought.
Options are "identical" for exact duplicates,
"offset" for duplicates with an fixed offset,
"multiply" where one duplicate is a multiple of another,
and "multiple_offset" where one duplicate is a multiple of another with an offset.}

\item{tolerance}{Small positive number to allow for numerical imprecision}

\item{reverse}{logical, test if the sequence is a reversed duplicate.}
}
\description{
Find and extract duplicate sequences of specified length
}
\details{
Note that type = "multiply_offsets" will also find multiplied duplicates (offset = 0),
constant offsets (multiplier = 1), and identical duplicates (multiplier = 1, offset = 0)
}
\examples{
set.seed(42)
x <- rnorm(10)
x <- c(x, x[1:4] + 2.03)
detectduplicate:::dup_find_length_n(vec = x, n = 5, type = "offset")
}
