% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dup_permute.R
\name{dup_permute}
\alias{dup_permute}
\title{Null expectation for apparent sequences}
\usage{
dup_permute(vec, min = 5, max = 10, nsamp = 100, type = "identical")
}
\arguments{
\item{vec}{vector with possible duplicates}

\item{min}{minimum length of sequence sought}

\item{max}{maximum length of sequence sought}

\item{nsamp}{number of replications}

\item{type}{Type of duplicate sought: "identical", "offset", "multiply" or
"multiply_offset"}
}
\description{
Vectors with low cardinality have a high risk of false positive
duplicate sequences, especially for short sequences.
\code{dup_permute} permutes sequences and then tests for duplicates.
}
\examples{
# data from https://doi.org/10.5061/dryad.bd26mq0
TotalPrey <- c(
  4, 5, 4, 1, 3, 8, 2, 4, 7, 6, 1, 8, 3, 6, 2, 1, 3, 5, 2, 7,
  1, 1, 3, 4, 4, 1, 3, 4, 3, 9, 2, 7, 6, 5, 3, 2, 4, 3, 4, 3, 2,
  3, 4, 3, 3, 3, 2, 4, 4, 2, 2, 6, 4, 3, 5, 2, 4, 3, 5, 1, 3, 4,
  4, 3, 1, 3, 2, 2, 3, 1, 4, 4, 3, 4, 3, 4, 3, 3, 7, 3, 4, 9, 2,
  7, 3, 5, 2, 3, 6, 2, 4, 5, 2, 8, 5, 3, 4, 5, 6, 4, 3, 3, 4, 3,
  2, 3, 2, 2, 2, 1, 1, 2, 1, 3, 4, 1, 2, 3, 3, 4, 1, 2, 3, 3, 2,
  2, 3, 1, 5, 3, 4, 2, 2, 1, 1, 2, 1, 0, 1, 2
)

res <- dup_permute(TotalPrey, min = 5, max = 10, nsamp = 100)
res
res |> dplyr::summarise(prob_has_duplicate = mean(sequences > 0), .by = "n")
}
