% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dup_plot.R
\name{dup_plot}
\alias{dup_plot}
\title{Plot duplicates}
\usage{
dup_plot(obj, x)
}
\arguments{
\item{obj}{result from \code{dup_find_all}}

\item{x}{The vector or data.frame used}
}
\description{
Plot duplicates
}
