% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dup_plot.R
\name{dup_plot_raster}
\alias{dup_plot_raster}
\title{Plot duplicates as a raster}
\usage{
dup_plot_raster(obj)
}
\arguments{
\item{obj}{result from \code{dup_find} or \code{dup_find_all}}
}
\description{
Plot duplicates as a raster
}
