% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_extract.R
\name{finv}
\alias{finv}
\title{1D index to 2D index}
\usage{
finv(k, dist_obj)
}
\arguments{
\item{k}{Indicies to convert to 2D}

\item{dist_obj}{distance object}
}
\description{
1D index to 2D index
}
\references{
https://stackoverflow.com/a/39006382/2055765
}
