% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kp2014.R
\docType{data}
\name{kp2014}
\alias{kp2014}
\title{Spider aggression data}
\format{
A \code{"tibble"}.
}
\source{
\href{https://figshare.com/articles/Data_for_Keiser_and_Pruitt_2014_-_Behavioral_Ecology/11778552}{figshare}
}
\usage{
data(kp2014)
}
\description{
Spider aggression data for Keiser and Pruitt 2014 - Behavioral Ecology.
}
\examples{
data(kp2014)
dup_find_all(
  x = kp2014$`Theridion murarium Aggressiveness...4`,
  type = "offset"
)
}
\references{
Keiser, C. N., Pruitt, J.N. Spider aggressiveness determines the
bidirectional consequences of host–inquiline interactions,
\emph{Behavioral Ecology}, 25: 142–151 2014,
(\href{http://dx.doi.org/10.1093/beheco/art096}{https://doi.org/10.1093/beheco/art096})
}
\keyword{datasets}
