% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/last_2_digits.R
\name{last2_diff_test}
\alias{last2_diff_test}
\title{last two digits}
\usage{
last2_diff_test(x, n = 100)
}
\arguments{
\item{x}{Vector of integer values (either integer or numeric class)}

\item{n}{Number of replications required}
}
\description{
Tests if the distribution of difference between the last two digits of a vector of numbers departs from the expected distribution
}
\examples{
last2_diff_test(sample(10:99, size = 1000, replace = TRUE), 100)
last2_diff_test(c(10:99, 1 + rep(c(11 * (1:8)), 2)), 100)

}
