% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{near_dist}
\alias{near_dist}
\alias{near_dist_cpp}
\title{Detect rows that are (within tolerance) identical}
\usage{
near_dist_cpp(k, tolerance, rev, negate)
}
\arguments{
\item{k}{Numeric vector}

\item{tolerance}{Numeric The tolerance for comparing values.}

\item{rev}{boolean Test for reversed duplicates e.g ABC matches CBA}

\item{negate}{boolean Also test for negative duplicates.}
}
\value{
A distance object containing TRUE for rows that are identical (within tolerance), false otherwise
}
\description{
This function takes a numeric matrix and runs a pairwise comparison
of rows, reporting whether they are identical or not, with a tolerance for small differences.
}
