% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment_cor.R
\name{segment_cor}
\alias{segment_cor}
\title{segment_cor}
\usage{
segment_cor(vec, n)
}
\arguments{
\item{vec}{numeric vector to test for possible offset duplicate sequences}

\item{n}{length of duplicate sequence to search for}
}
\description{
segment_cor
}
\examples{
set.seed(42) 
x <- rnorm(10)
x <- c(x, x[1:4])
segment_cor(vec = x, n = 4)
}
