% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment_thresh.R
\name{segment_thresh}
\alias{segment_thresh}
\title{Threshold segment correlations}
\usage{
segment_thresh(seg, threshold, tolerance = sqrt(.Machine$double.eps))
}
\arguments{
\item{seg}{Output of \code{\link[=segment_cor]{segment_cor()}}}

\item{threshold}{Minimum correlation to accept. Defaults to 1, possible range -1 to 1 (but values much below 1 make no sense)}

\item{tolerance}{Small positive number to allow for numerical imprecision}
}
\description{
Thresholds correlation between all segments length n of a vector
}
\details{
Tests if correlations are >= than a given threshold.
}
\examples{
set.seed(42) 
x <- rnorm(10)
x <- c(x, x[1:4])
seg <- segment_cor(vec = x, n = 4)
segment_thresh(seg)    
}
