% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_cor.R
\name{sequence_cor}
\alias{sequence_cor}
\title{segment_cor}
\usage{
sequence_cor(vec, n)
}
\arguments{
\item{vec}{numeric vector to test for possible offset duplicate sequences}

\item{n}{length of duplicate sequence to search for}
}
\description{
Finds correlation between all segments length n of a vector
}
\details{
Embeds the vector into low-dimensional Euclidean space and then finds correlation between all columns. Pairs of columns with a correlation of 1 are either duplicates or have an offset.
}
\examples{
set.seed(42)
x <- rnorm(10)
x <- c(x, x[1:4] + 2.03)
sequence_cor(vec = x, n = 4)
}
