% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_dist.R
\name{sequence_dist}
\alias{sequence_dist}
\title{segment_dist}
\usage{
sequence_dist(vec, n, type = c("identical", "offset", "multiple"), tolerance)
}
\arguments{
\item{vec}{numeric vector to test for possible offset duplicate sequences}

\item{n}{length of duplicate sequence to search for}

\item{type}{Type of duplicate sought. See \code{\link[=sequence_find]{sequence_find()}} for details.}

\item{tolerance}{Small positive number to allow for numerical imprecision}
}
\description{
Finds distance between all segments length n of a vector
}
\details{
Embeds the vector into low-dimensional Euclidean space and then
finds manhattan distance between all pairs of rows
Pairs of rows with a distance of 0 are identical if \code{type} is "identical".
If \code{type} is "offset, rows are de-meaned before the distance is calculated.
If \code{type} is multiple, rows are standardised to zero mean and unit
standard deviation before the distance is calculated.
}
\examples{
set.seed(42)
x <- rnorm(10)
x <- c(x, x[1:4] + 2.03)
twopointzerothree:::sequence_dist(vec = x, n = 4, type = "offset")
}
