% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_find.R
\name{sequence_find}
\alias{sequence_find}
\title{Find and extract duplicate sequences}
\usage{
sequence_find(vec, n, type, tolerance)
}
\arguments{
\item{vec}{vector with possible duplicates}

\item{n}{length of sequence sought}

\item{type}{Type of duplicate sought. See details.}

\item{tolerance}{Small positive number to allow for numerical imprecision}
}
\description{
Find and extract duplicate sequences
}
\details{
Options for \code{type} are "identical" for exact duplicates (new == old),
"offset" for duplicates with an fixed offset (new == old + constant ),
and "multiple" for duplicates with a multiplicative offset,
and perhaps a constant offset (new == old * constant1 + constant2).

Note that multiplicative offsets will also find constant offsets
(i.e. constant1 = 1), and identical duplicates (i.e. constant2 = 0)
}
\examples{
set.seed(42)
x <- rnorm(10)
x <- c(x, x[1:4] + 2.03)
twopointzerothree:::sequence_find(vec = x, n = 5, type = "offset")
}
