% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_longest.R
\name{sequence_longest}
\alias{sequence_longest}
\title{Find longest duplicate sequence}
\usage{
sequence_longest(vec, min = 3, max = NULL, ...)
}
\arguments{
\item{vec}{Numeric vector to test for (offset) duplicate values}

\item{min}{Minimum length sequence to search for.}

\item{max}{Maximum length sequence to search for. If \code{NULL}, finds longest possible.}

\item{...}{arguments to \link{segment_thresh}}
}
\description{
Finds the longest sequence of (offset) duplicate values
}
\details{
Loops to find the longest possible sequence of (offset) duplicate values.
}
\examples{
set.seed(42) 
x <- rnorm(10)
x <- c(x, x[1:4] + 2.03)
sequence_longest(vec = x)    
}
