% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_thresh.R
\name{sequence_thresh}
\alias{sequence_thresh}
\title{Threshold segment correlations}
\usage{
sequence_thresh(sequ, threshold = 1, tolerance = sqrt(.Machine$double.eps))
}
\arguments{
\item{sequ}{Output of \code{\link[=sequence_cor]{sequence_cor()}}}

\item{threshold}{Minimum correlation to accept. Defaults to 1, possible range -1 to 1 (but values much below 1 make no sense)}

\item{tolerance}{Small positive number to allow for numerical imprecision}
}
\description{
Thresholds correlation between all segments length n of a vector
}
\details{
Tests if correlations are >= than a given threshold (with tolerance for imprecision in correlations). Default threshold is 1 (perfect correlation). Might be useful to reduce this slightly, or equivalently increase tolerance to allow for cases where imprecision is high.
}
\examples{
set.seed(42)
x <- rnorm(10)
x <- c(x, x[1:4] + 2.03)
sequ <- sequence_cor(vec = x, n = 4)
sequence_thresh(sequ)
}
