% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dup_dist.R
\name{dup_dist}
\alias{dup_dist}
\title{segment_dist}
\usage{
dup_dist(
  vec,
  n,
  type = c("identical", "offset", "multiply", "multiply_offset"),
  tolerance,
  reverse = FALSE
)
}
\arguments{
\item{vec}{numeric vector to test for possible offset duplicate sequences}

\item{n}{length of duplicate sequence to search for}

\item{type}{Type of duplicate sought. See \code{\link[=dup_find]{dup_find()}} for details.}

\item{tolerance}{Small positive number to allow for numerical imprecision}

\item{reverse}{logical, test if the sequence is a reversed duplicate.}
}
\description{
Finds distance between all segments length n of a vector
}
\details{
Embeds the vector into low-dimensional Euclidean space and then
tests is rows are identical within tolerance.
Pairs of rows with a distance of 0 are identical if \code{type} is "identical".
If \code{type} is "offset, rows are centred before the distance is calculated.
If \code{type} is multiply, rows are scaled without centring.
If \code{type} is multiply_offset, rows are centred and scales.
}
\examples{
set.seed(42)
x <- rnorm(10)
x <- c(x, x[1:4] + 2.03)
detectduplicate:::dup_dist(vec = x, n = 4, type = "offset")
}
