% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dup_find_all.R
\name{dup_find_all}
\alias{dup_find_all}
\alias{dup_find_all.default}
\alias{dup_find_all.data.frame}
\title{Find and extract all duplicate sequences}
\usage{
dup_find_all(x, type = "identical", min = 4, tolerance, reverse)

\method{dup_find_all}{default}(x, type = "identical", min = 4, tolerance, reverse)

\method{dup_find_all}{data.frame}(x, type = "identical", min = 4, tolerance, reverse)
}
\arguments{
\item{x}{vector or data.frame with possible duplicates}

\item{type}{Type of duplicate sought.
Options are "identical" for exact duplicates,
"offset" for duplicates with an fixed offset,
"multiply" where one duplicate is a multiple of another,
and "multiple_offset" where one duplicate is a multiple of another with an offset.}

\item{min}{length of the shortest sequence of interest
(high risk of false positives if this is short)}

\item{tolerance}{Small positive number to allow for numerical imprecision.
Defaults to \code{sqrt(.Machine$double.eps)}.}

\item{reverse}{logical, test if the sequence is a reversed duplicate.
Defaults to FALSE}
}
\description{
Find and extract all duplicate sequences
}
\examples{
data(kp2014)
dup_find_all(
  x = kp2014$`Theridion murarium Aggressiveness...4`,
  type = "offset"
)
dup_find_all(
  x = kp2014[, c("Theridion murarium Aggressiveness...4",
                 "Theridion murarium Aggressiveness...5",
                 "Theridion murarium Aggressiveness...6")],
  type = "offset"
)
}
