% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment_cor.R
\name{segment_cor}
\alias{segment_cor}
\title{segment_cor}
\usage{
segment_cor(vec, n)
}
\arguments{
\item{vec}{numeric vector to test for possible offset duplicate sequences}

\item{n}{length of duplicate sequence to search for}
}
\description{
Finds correlation between all segments length n of a vector
}
\details{
Embeds the vector into low-dimensional Euclidean space and then finds correlation between all columns. Pairs of columns with a correlation of 1 are either duplicates or have an offset.
}
\examples{
set.seed(42) 
x <- rnorm(10)
x <- c(x, x[1:4])
segment_cor(vec = x, n = 4)
}
