% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_extract.R
\name{sequence_extract}
\alias{sequence_extract}
\title{extract duplicate sequences}
\usage{
sequence_extract(th, vec, n)
}
\arguments{
\item{th}{result of sequence_thresh}

\item{vec}{vector with possible duplicates}

\item{n}{length of sequence sought}
}
\description{
extract duplicate sequences
}
\examples{
#' set.seed(42)
x <- rnorm(10)
x <- c(x, x[1:4] + 2.03)
d <- twopointzerothree:::sequence_dist(vec = x, n = 3)
twopointzerothree:::sequence_extract(d, vec = x)
}
