% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_find.R
\name{sequence_find}
\alias{sequence_find}
\title{Find and extract duplicate sequences}
\usage{
sequence_find(vec, n)
}
\arguments{
\item{vec}{vector with possible duplicates}

\item{n}{length of sequence sought}
}
\description{
Find and extract duplicate sequences
}
\examples{
set.seed(42)
x <- rnorm(10)
x <- c(x, x[1:4] + 2.03)
sequence_find(vec = x, n = 5)
}
