% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_thresh.R
\name{sequence_thresh}
\alias{sequence_thresh}
\title{Threshold inter-segment distances}
\usage{
sequence_thresh(dist, tolerance)
}
\arguments{
\item{dist}{Output of \code{\link[=sequence_dist]{sequence_dist()}}}

\item{tolerance}{Small positive number to allow for numerical imprecision}
}
\description{
Thresholds distances between all segments length n of a vector
}
\details{
Tests if distances are, within tolerance, equal to zero.
}
\examples{
set.seed(42)
x <- rnorm(10)
x <- c(x, x[1:4] + 2.03)
sequ <- twopointzerothree:::sequence_dist(vec = x, n = 4)
twopointzerothree:::sequence_thresh(sequ)
}
